practical_task4/flask_app/requirements.txt
flask
azure-keyvault-secrets
azure-identity